app/pyproject.toml
[tool.poetry]
name = "auto-index"
version = "0.1.0"
description = "A sample Python project"
authors = ["Viet Do<vietdqhcmute@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
python-docx = "^1.1.2"


[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
